+++ CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
 
-
+## [6.2.3] - 2019-04-11
+- Fix serviceContext getting removed on update service instance, when no context is set
+- Fix serviceInstances which failed during provision count as active
+- Add publishing to bintray
 
 ## [6.2.0] - 2019-03-28
 - [MINOR] Refactor to separate broker-core from broker